Version 1.2 - Inspector, Game Objects, & Gizmos
Two windows have been added, the inspector, which allows for properties of objects to be edited, and the adding window, which only adds blank game objects at the moment. World grid has been added. Game objects are now displayed in the world window. Game object position and size can be altered in the world with gizmos. As always, the next step is to abstract these changes, as they have been lazily added.